requirements.txt
numpy
wandb
gitpython
opencv-python
joblib
libigl
hydra-core
hydra-zen
imageio
open3d
pygltflib
rich
inflection
plyfile
polyscope
robust-laplacian
pymeshlab
deltaconv
pygeodesic
siren-pytorch
lightly
triangle
pyglet
pyfmaps
seaborn
scikit-learn
noise
lightning
torch_geometric
scipy
jupyterlab
pandas
meshio
trimesh
cmake
lz4
boost
pybind11
pytest
torch-householder
kornia